Modify structure of huron_driver
